(FM-6769) rewrite the unit tests for NetworkInterface and other cleanups (#63)
* Allow symbols in safely loaded YAMLs

This is disabled by default because malicious document input can use
symbols for a memory exhaustion attack. Since we're only reading local
trusted files, this is not a high issue, and we'll need symbols in the
test data.

* Avoid polluting the global namespace

By removing this `include`, the global namespace doesn't get polluted
with a bunch of rspec methods.

This will need to be replicated to the other spec files.

* Sort requires alphabetically

* Refactor the NetworkInterface provider

* cache the commands_hash once at the class level
* lift the `parse` and `command` methods to the class level

* (FM-6769) rewrite the unit tests for NetworkInterface

These tests closely follow the NetworkInterface provider's structure.
The first few lines now test the `interface_parse_out` and
`interface_config_command` functions directly. Since they are now class methods
(see previos commit), they are independent of any state of provider instances,
and testing them is easily achieved by comparing against the test data from the
YAML file.

The YAML file has been changed to contain the actual data structures. Thanks to
allowing symbols in `load_yaml`, this now works fine. Through this, we can
avoid the use of `eval` here.

The second half of the tests only has to superficially then test that the main
methods of the provider wire up the utility methods correctly, and hand off the
right commands to the backend. This feels like it could als be lifted into the data file, but references to provider specific methods (like
`run_command_interface_mode`) make that more complicated than expected.

* Partially update to 1.4.0.pre template

* Fix rubocop