(FM-7029) Fixup network_dns This commit fixes a couple of issues with network_dns First is the change in CLI from `domain-name` / `domain-list` to `domain name` / `domain list` Starting in IOS 12.2 the commands were moved to the later format.
Although systems still support setting via the old style, and even report in it - it is no longer preferred.
Addtionally most systems now report the value in new style.

This commit takes that into account by reading either format, and only setting in the new one.

The second is the handling of name-server's specified on a single line.
At some point IOS moved from multiple single value entries to a single multi-value entry

```
ip name-server 1.1.1.1
ip name-server 1.0.0.1
ip name-server 192.168.198.1

ip name-server 192.168.198.1 1.1.1.1 1.0.0.1
```

This commit takes those values into account.  Some of the magic there is handled in the provider itself.
If there is a better way to handle this directly in the regex that would be even more better :)