(FM-6990) ios_config idempotent via master and regex options (#189)
This commit adds two major sets of functionality for ios_config
The first is idempotency when being run from a master.  We use the canonicalize feature of RSAPI to execute idempotent_regex when set to determine if resource is already in desired state.
The second is regex options via idempotent_regex_options.  This allows users to specify regex options such as ignorecase and multiline.