(maint) switch the Device to url_or_config handling
The Resource API allows a `Hash` with the credentials/config instead
of a `file://` URL [since 1.4.0](https://github.com/puppetlabs/puppet-resource_api/pull/96).

This commit switches over to using the Resource API Device baseclass
to handle this behaviour for us. It also avoids the undeclared dependency
on the hocon gem.