(maint) updating the error and warning checks
Previously it was only checking line starts with `Error:` or `Warning:` - but colour output in the device runs meant the line started with a colour hex.

This change will highlight areas of the test which are failing across devices.

Also updating the MTU value as some older devices has a lowest bound of 1501.