(maint) Fix for NTP Server test
Now that we execute locally, we pick up all debug.
One of our test fixtures was picking up all debug output, but we wished to parse only the Puppet device output as yaml.
This commit allows for the verbose, debug, trace option string to be set to an empty string (ie. not set) if desired.