(maint) workaround faulty change detection logic
The original regular expression `%r{^Notice: /Stage\[main\]}` does not
actually match change events, as there is a hidden control character
at the start of the line. The fixes in the "(maint) improve `run_device`
method" change uncloaked that.

This commit changes all violations of that into an error message for
later cleanup.