(FM-8310) idempotency fixes for the ios network trunk resource
The resource allowed you to specify things like `['add', '1-10']` for allowed vlans but the CLI would always return the vlans as a string which caused idempotency issues.

Our options would have been to force users to provide us with the allowed vlans in the format expected, i.e. '1,10-12,20-100' which wouldn't be fair as using the cli allows them to do things like `except 10-20` which meant they don't need to know the max/min allowed vlans etc.

So to keep idempotency the use of the canonicalize has been used to take the values and convert them to the expected cli output, i.e. `['add', '1-10']` will get the current and add these values, for example:

```
current = '100,250-256'
['add', '1-10']
result = '1-10,100,250-256'
```