(FM-8310) idempotency fixes for ios radius_global
The `attributes` property is something that can be used too manage attributes - `[]` is used to remove attributes, when attributes are removed then of course the current_state will not return anything.

So as a compromise the `attributes` will only be sent as `[]` if there is currently attributes set on the device, so if the attributes are removed from the device and `[]` is in the manifest then the canonicalize will remove the `attributes` property as it is not required as there is no attributes on the device.