(FM-8310) idempotency fixes for tacacs_server
Key_format is assumed 0 if there is a key set without a key_format being returned from the device

Hostname from the device returns it in a very specific compressed format when using IPv6 addresses. This change will compress the hostname when it is IPv6 to match the values returned from the device.

Key, key_format, and timeout will be returned with defaults, `'unset'` for key and 0 for key_format and timeout to signify that these are not managed on the device.