(maint) remove rsapi<1.3 compatibility code
Based on the comment in https://github.com/puppetlabs/cisco_ios/pull/181#pullrequestreview-122989646 these entries were only for compatibility with Resource API prior to 1.3 and can therefore be removed now.

The extra loading of the legacy type and provider was causing codecoverage to break due to double-loading of the provider implementation files.