(FM-7815) Merge ios_access_list and ios_acl_entry into ios_acl
Prior to this work `ios_access_list` allowed for a named access list
to be created without function, and it was not possible to create a
numbered access list.

This work merges `ios_access_list` and `ios_acl_entry` into one type:
`ios_acl`, which allows for a simpler and more robust implementation,
while at the same time reducing duplication.

Resources will use the title_pattern of
`<access_list> <access_list_type> <entry>`. For example,
`RepelInvaders extended 10` will create an entry in the `extended`
access list called `RepelInvaders` with the priority entry `10`.
While `10 standard 10` will create an entry in the numbered access list 10
as standard with a priority of 10.