(PE-29241) Use all supported algorithms for ssh connections
Some older, less secure algorithms have been deprecated with net-ssh 6.x (which is what is available in the context the ace server in PE executes remote tasks). This commit adds an option to net-ssh to include the less secure algorithms which are on the path to deprecation until they are formally deprecated/removed.